Minior github-http wrapper refactoring.
All of the methods in the library are now part of the class. There is
now only one get() method for performing HTTP requests. The previous
code had a thin get() wrapper around the actual get() code. This made
reading the code more complex. The code also had multiple paths to
cleaning up http responses. The new code should have all that reduced
down to one place, in the get() method that generates the response.